requirements.txt
blessed>=1.20.0
asyncio
websockets
pyyaml
